Defining the Hydra Way for Curate
The Hydra Stack has a set of promises that we should call greater
attention to. This is especially important for new developers.

We have an anti-harassment policy that has been defined and ratified.

We have a set of community principles that we should call attention
to.

Also removing Curate references from CONTRIBUTING.md in preparation
for community adoption of the Contributing document into Hydra's core
gems.

Related to projecthydra/hydra#34
Related to projecthydra/hydra#35
Related to projecthydra/hydra#36
[skip ci]